Some changes in public

+++ public/UtlCachedFileData.h
@@ -23,7 +23,7 @@
 // If you change to serialization protocols, this must be bumped...
 #define UTL_CACHE_SYSTEM_VERSION 2
 
-#define UTL_CACHED_FILE_DATA_UNDEFINED_DISKINFO (long)-2
+#define UTL_CACHED_FILE_DATA_UNDEFINED_DISKINFO -2
 
 // Cacheable types must derive from this and implement the appropriate methods...
 abstract_class IBaseCacheInfo
@@ -102,7 +102,7 @@ public:
 		return idx != m_Elements.InvalidIndex() ? true : false;
 	}
 
-	void SetElement(char const *name, long fileinfo, T *src)
+	void SetElement(char const *name, int fileinfo, T *src)
 	{
 		SetDirty(true);
 
@@ -149,7 +149,7 @@ public:
 	void ForceRecheckDiskInfo();
 	// Iterates all entries and gets filesystem info and optionally causes rebuild on any existing items which are out
 	// of date
-	void CheckDiskInfo(bool force_rebuild, long cacheFileTime = 0L);
+	void CheckDiskInfo(bool force_rebuild, int cacheFileTime = 0);
 
 	void SaveManifest();
 	bool ManifestExists();
@@ -159,7 +159,7 @@ public:
 		return m_sRepositoryFileName;
 	}
 
-	long GetFileInfo(char const *filename)
+	int GetFileInfo(char const *filename)
 	{
 		ElementType_t element;
 		element.handle = g_pFullFileSystem->FindOrAddFileName(filename);
@@ -221,8 +221,8 @@ private:
 		}
 
 		FileNameHandle_t handle;
-		long fileinfo;
-		long diskfileinfo;
+		int fileinfo;
+		int diskfileinfo;
 		int dataIndex;
 	};
 
@@ -256,7 +256,7 @@ T *CUtlCachedFileData<T>::Get(char const *filename)
 	{
 		e.fileinfo = 0;
 	}
-	long cachefileinfo = e.fileinfo;
+	int cachefileinfo = e.fileinfo;
 	// Set the disk fileinfo the first time we encounter the filename
 	if(e.diskfileinfo == UTL_CACHED_FILE_DATA_UNDEFINED_DISKINFO)
 	{
@@ -813,7 +813,7 @@ T *CUtlCachedFileData<T>::RebuildItem(const char *filename)
 
 	ForceRecheckDiskInfo();
 
-	long cachefileinfo = e.fileinfo;
+	int cachefileinfo = e.fileinfo;
 	// Set the disk fileinfo the first time we encounter the filename
 	if(e.diskfileinfo == UTL_CACHED_FILE_DATA_UNDEFINED_DISKINFO)
 	{
@@ -894,7 +894,7 @@ public:
 
 // Iterates all entries and causes rebuild on any existing items which are out of date
 template<class T>
-void CUtlCachedFileData<T>::CheckDiskInfo(bool forcerebuild, long cacheFileTime)
+void CUtlCachedFileData<T>::CheckDiskInfo(bool forcerebuild, int cacheFileTime)
 {
 	char fn[512];
 	int i;
