+++ public/bspfile.h
@@ -682,7 +682,7 @@ public:
 	{
 		ALLOWEDVERTS_SIZE = PAD_NUMBER(MAX_DISPVERTS, 32) / 32
 	};
-	unsigned long m_AllowedVerts[ALLOWEDVERTS_SIZE]; // This is built based on the layout and sizes of our neighbors
+	unsigned int m_AllowedVerts[ALLOWEDVERTS_SIZE]; // This is built based on the layout and sizes of our neighbors
 													 // and tells us which vertices are allowed to be active.
 };
