+++ public/dt_send.cpp
@@ -244,7 +244,7 @@ void SendProxy_UInt16ToInt32(const SendProp *pProp, const void *pStruct, const v
 void SendProxy_UInt32ToInt32(const SendProp *pProp, const void *pStruct, const void *pData, DVariant *pOut,
 							 int iElement, int objectID)
 {
-	*((unsigned long *)&pOut->m_Int) = *((unsigned long *)pData);
+	*((unsigned int *)&pOut->m_Int) = *((unsigned int *)pData);
 }
 #ifdef SUPPORTS_INT64
 void SendProxy_UInt64ToInt64(const SendProp *pProp, const void *pStruct, const void *pData, DVariant *pOut,
@@ -298,18 +298,18 @@ void *SendProxy_SendLocalDataTable(const SendProp *pProp, const void *pStruct, c
 // ---------------------------------------------------------------------- //
 float AssignRangeMultiplier(int nBits, double range)
 {
-	unsigned long iHighValue;
+	unsigned int iHighValue;
 	if(nBits == 32)
 		iHighValue = 0xFFFFFFFE;
 	else
-		iHighValue = ((1 << (unsigned long)nBits) - 1);
+		iHighValue = ((1 << (unsigned int)nBits) - 1);
 
 	float fHighLowMul = iHighValue / range;
 	if(CloseEnough(range, 0))
 		fHighLowMul = iHighValue;
 
 	// If the precision is messing us up, then adjust it so it won't.
-	if((unsigned long)(fHighLowMul * range) > iHighValue || (fHighLowMul * range) > (double)iHighValue)
+	if((unsigned int)(fHighLowMul * range) > iHighValue || (fHighLowMul * range) > (double)iHighValue)
 	{
 		// Squeeze it down smaller and smaller until it's going to produce an integer
 		// in the valid range when given the highest value.
@@ -318,7 +318,7 @@ float AssignRangeMultiplier(int nBits, double range)
 		for(i = 0; i < ARRAYSIZE(multipliers); i++)
 		{
 			fHighLowMul = (float)(iHighValue / range) * multipliers[i];
-			if((unsigned long)(fHighLowMul * range) > iHighValue || (fHighLowMul * range) > (double)iHighValue)
+			if((unsigned int)(fHighLowMul * range) > iHighValue || (fHighLowMul * range) > (double)iHighValue)
 			{
 			}
 			else
