+++ public/tier2/riff.h
@@ -24,12 +24,12 @@
 class IFileReadBinary
 {
 public:
-	virtual int open(const char *pFileName) = 0;
-	virtual int read(void *pOutput, int size, int file) = 0;
-	virtual void close(int file) = 0;
-	virtual void seek(int file, int pos) = 0;
-	virtual unsigned int tell(int file) = 0;
-	virtual unsigned int size(int file) = 0;
+	virtual intp open(const char *pFileName) = 0;
+	virtual int read(void *pOutput, int size, intp file) = 0;
+	virtual void close(intp file) = 0;
+	virtual void seek(intp file, int pos) = 0;
+	virtual unsigned int tell(intp file) = 0;
+	virtual unsigned int size(intp file) = 0;
 };
 
 //-----------------------------------------------------------------------------
@@ -108,11 +108,11 @@ private:
 class IFileWriteBinary
 {
 public:
-	virtual int create(const char *pFileName) = 0;
-	virtual int write(void *pData, int size, int file) = 0;
-	virtual void close(int file) = 0;
-	virtual void seek(int file, int pos) = 0;
-	virtual unsigned int tell(int file) = 0;
+	virtual intp create(const char *pFileName) = 0;
+	virtual int write(void *pData, int size, intp file) = 0;
+	virtual void close(intp file) = 0;
+	virtual void seek(intp file, int pos) = 0;
+	virtual unsigned int tell(intp file) = 0;
 };
 //-----------------------------------------------------------------------------
 // Purpose: Used to write a RIFF format file
